Rewrite listr.py introducing 'request objects' & removing formatting
The whole listr.py file was rewritten to simplify what it does.  Now, it is no
longer concerned with sorting or formatting the results.  It merely gathers data
and then hands it off to a presentation layer which handles all formatting and
organization.

To facilitate this tightening of scope, the main data gathering pipeline now
works like this:
  1. Retrieve card requests from the user (now supports quantities and printing
  preferences)
  2. Get the card data associated with the requested cards from all_cards.json
  3. Merge select card data with the request object to create an enriched
  request
  4. Pass enriched requests off to the appropriate presentation layer